Refactor Article, Authentication, Submission to sub application and add prefix /api to them
