+++ .env.example
@@ -1,7 +1,7 @@
 APPLICATION_SECRET=
 TWITTER_APP_KEY=
 TWITTER_APP_SECRET=
-TWITTER_CALLBACK_URL=http://localhost:8080/auth/twitter/callback
+TWITTER_CALLBACK_URL=http://localhost:8080/api/auth/twitter/callback
 DATABASE_URL=postgres://username:password@localhost:5432/analog_development
 MONGODB_URI=mongodb://localhost:27017/analogCafeDev
 REDIS_URL=redis://localhost:6379
