+++ app/authentication/ability.js
@@ -1,61 +1,63 @@
-const UrlPattern = require('url-pattern');
+const UrlPattern = require('url-pattern')
 
 var permissions = [
   {
-    method: "get",
-    url: "/",
-    roles: ["member", "admin", "guest"]
+    method: 'get',
+    url: '/api',
+    roles: ['member', 'admin', 'guest']
   },
   {
-    method: "get",
-    url: "/me",
-    roles: ["member", "admin"]
+    method: 'get',
+    url: '/api/me',
+    roles: ['member', 'admin']
   },
   {
-    method: "get",
-    url: "/submissions",
-    roles: ["member", "admin", "guest"]
+    method: 'get',
+    url: '/api/submissions',
+    roles: ['member', 'admin', 'guest']
   },
   {
-    method: "post",
-    url: "/submissions",
-    roles: ["member", "admin"]
+    method: 'post',
+    url: '/api/submissions',
+    roles: ['member', 'admin']
   },
   {
-    method: "get",
-    url: "/submissions/:submissionId",
-    roles: ["member", "admin", "guest"]
+    method: 'get',
+    url: '/api/submissions/:submissionId',
+    roles: ['member', 'admin', 'guest']
   },
   {
-    method: "put",
-    url: "/submissions/:submissionId",
-    roles: ["member", "admin"]
+    method: 'put',
+    url: '/api/submissions/:submissionId',
+    roles: ['member', 'admin']
   },
   {
-    method: "get",
-    url: "/auth/twitter",
-    roles: ["member", "admin", "guest"]
+    method: 'get',
+    url: '/api/auth/twitter',
+    roles: ['member', 'admin', 'guest']
   },
   {
-    method: "get",
-    url: "/auth/twitter/callback",
-    roles: ["member", "admin", "guest"]
-  },
+    method: 'get',
+    url: '/api/auth/twitter/callback',
+    roles: ['member', 'admin', 'guest']
+  }
 ]
 
-function cleanUrl(url){
-  return url.split("?")[0]
+function cleanUrl (url) {
+  return url.split('?')[0]
 }
 
-function isUserHasPermission(req){
-  for(var i = 0; i < permissions.length; i++ ){
-    var pattern = new UrlPattern(permissions[i]["url"])
+function isUserHasPermission (req) {
+  for (var i = 0; i < permissions.length; i++) {
+    var pattern = new UrlPattern(permissions[i]['url'])
     var isMatchUrl = pattern.match(cleanUrl(req.url)) != null
-    var isMatchMethod = permissions[i]["method"] == req.method.toLowerCase()
-    if(isMatchUrl && isMatchMethod){
-      var isMatchRoleGuest = !req.user && permissions[i]["roles"].indexOf("guest") > -1
-      var isMatchRole = req.user && permissions[i]["roles"].indexOf(req.user.role) > -1
-      if(isMatchRoleGuest || isMatchRole){
+    var isMatchMethod = permissions[i]['method'] === req.method.toLowerCase()
+    if (isMatchUrl && isMatchMethod) {
+      var isMatchRoleGuest =
+        !req.user && permissions[i]['roles'].indexOf('guest') > -1
+      var isMatchRole =
+        req.user && permissions[i]['roles'].indexOf(req.user.role) > -1
+      if (isMatchRoleGuest || isMatchRole) {
         return true
       }
     }
@@ -64,10 +66,10 @@ function isUserHasPermission(req){
 }
 
 function checkUserPermission (req, res, next) {
-  if(isUserHasPermission(req)){
-    next();
-  }else{
-    res.status(401).json({ message: 'You have no access.' });
+  if (isUserHasPermission(req)) {
+    next()
+  } else {
+    res.status(401).json({ message: 'You have no access.' })
   }
 }
 module.exports = checkUserPermission
